Move printing-related tests to their own test module
